moment bug fix (#235)
* bug fix

* i18n actions and reducer

* CHANGELOG.md update
