+++ src/actions.js
@@ -12,5 +12,3 @@ export const SET_DEV_CONFIG = "asab/setDevConfig";
 export const SET_SIDEBAR = "asab/setSidebar";
 export const CHANGE_SIDEBAR_SIZE = "asab/changeSidebarSize";
 export const SET_SMALL_SIDEBAR = "asab/setSmallSidebar"
-
-export const CHANGE_LANGUAGE = "asab/changeLanguage";
