+++ src/modules/i18n/actions.js
@@ -1,0 +1,1 @@
+export const CHANGE_LANGUAGE = "asab/changeLanguage";
