+++ src/modules/i18n/index.js
@@ -1,6 +1,7 @@
 import Module from 'asab-webui/abc/Module';
 import I18nService from './service.js';
 import LanguageDropdown from './dropdown.js';
+import languageReducer from './reducer.js'
 
 export default class I18nModule extends Module {
 
@@ -9,6 +10,7 @@ export default class I18nModule extends Module {
 		this.App = app;
 
 		this.I18nService = new I18nService(app, "I18nModule");
+		this.App.ReduxService.addReducer("language", languageReducer);
 	}
 
 	initialize() {
