--- src/reducers/languageReducer.js
@@ -1,14 +1,0 @@
-import { CHANGE_LANGUAGE } from "../actions";
-
-const initialValue = {
-    language: 'en'
-}
-
-export default (state = initialValue, action) => {
-    switch (action.type) {
-        case CHANGE_LANGUAGE:
-            return { ...state, language: action.language }
-        default:
-            return state
-    }
-}
